// src/types/element.ts
export type PropertiesGroup =
    'basic'
    | 'physical'
    | 'atomic'
    | 'other'
    | 'history'


export type Property =
    'symbol'
    | 'atomicNumber'
    | 'enName'
    | 'zhName'
    | 'atomicWeight'
    | 'pinyin'
    | 'group'
    | 'period'
    | 'block'
    | 'electronConfiguration'
    | 'electronsPerShell'
    | 'phaseAtSTP'
    | 'meltingPoint'
    | 'boilingPoint'
    | 'density'
    | 'heatOfFusion'
    | 'molarHeatCapacity'
    | 'oxidationStates'
    | 'electronegativity'
    | 'ionizationEnergies'
    | 'atomicRadius'
    | 'covalentRadius'
    | 'vanDerWaals'
    | 'spectralLines'
    | 'crystalStructure'
    | 'magneticOrdering'
    | 'speedOfSound'
    | 'thermalConductivity'
    | 'CASNumber'
    | 'discovery'
    | 'namedBy'


export type DisplayProperty = 'atomicWeight'
    | 'pinyin'
    | 'zhCNName&pinyin'
    | 'electronegativity'
    | 'atomicRadius'
    | 'meltingPoint'
    | 'boilingPoint'
    | 'electronConfiguration'

export type ColorSign = 'trend' | 'classification' | 'block' | 'state'

export type Emphasize = 'symbol' | 'name' | 'number'